+++ FindingAlice/Assets/Scenes/LoadingScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 141099594a0772a47a307d6979e35e69
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
