+++ FindingAlice/Assets/_Scripts/Clock/ClockTouchAndDrag.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0c5a9310af59cea42bc4424054660f7d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
