+++ FindingAlice/Assets/_Scripts/Clock/RangeCircle.cs
@@ -1,0 +1,14 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class RangeCircle : MonoBehaviour
+{
+    float angle;
+
+    void Update()
+    {
+        angle += Time.deltaTime * 500f;
+        transform.rotation = Quaternion.Euler(0, 0, angle);
+    }
+}
