+++ FindingAlice/Assets/_Scripts/Clock/WatchBand.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0649690276d69f4cafb7765c37cbf17
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
