+++ FindingAlice/Assets/_Scripts/GameSceneChange.cs
@@ -11,10 +11,13 @@ public class GameSceneChange : MonoBehaviour
         if(DataController.Instance._gameData.playerPosition != new Vector3(0,0,0))
         {
             Debug.Log("데이터 소멸 경고");
-            SceneManager.LoadScene("GameScene");
+
+            AsyncLoading.LoadScene("GameScene");
+            //SceneManager.LoadScene("GameScene");
         }
         else
-        SceneManager.LoadScene("GameScene");
+        AsyncLoading.LoadScene("GameScene");
+        //SceneManager.LoadScene("GameScene");
     }
 
     public void LoadGame(){
@@ -22,7 +25,9 @@ public class GameSceneChange : MonoBehaviour
         {
             return;
         }
-            checkLoad = true;
-            SceneManager.LoadScene("GameScene");
+        checkLoad = true;
+
+        AsyncLoading.LoadScene("GameScene");
+        //SceneManager.LoadScene("GameScene");
     }
 }
