+++ FindingAlice/Assets/_Scripts/LoadingScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a3b44df02d0b79e469c19739c4f07337
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
