+++ FindingAlice/Assets/_Scripts/LoadingScene/AsyncLoading.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 930c7c389cb5f6549bdb394ca7e14d37
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
