+++ FindingAlice/Assets/_Scripts/LoadingScene/NiddleRotate.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b60d29861e381bd439f0efa55693b465
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
