+++ FindingAlice/Assets/_Scripts/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cfb6e611d1d19db40b6dfb9f8e4d841c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
