+++ FindingAlice/Assets/_Scripts/Player/Check_CollisionToWall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30e43770eb3647b498dc7eeaec078f78
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
