--- FindingAlice/Assets/_Scripts/RangeCircle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d2745fe9baa2ea940a864b183db4dacc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
