--- FindingAlice/Assets/_Scripts/WatchBand.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a0649690276d69f4cafb7765c37cbf17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
