+++ FindingAlice/ProjectSettings/EditorBuildSettings.asset
@@ -17,4 +17,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/TutorialScene.unity
     guid: f3e5530aa47796f44a8952a435426e4a
+  - enabled: 1
+    path: Assets/Scenes/LoadingScene.unity
+    guid: 141099594a0772a47a307d6979e35e69
   m_configObjects: {}
